lockdep: be nice about building from userspace
Lockdep is an awesome piece of code which detects locking issues which are
relevant both to userspace and kernelspace. We can easily make lockdep
work in userspace since there is really no kernel spacific magic going
on in the code.

All we need is to wrap two functions which are used by lockdep and are
very kernel specific.

Doing that will allow tools located in tools/ to easily utilize lockdep's
code for their own use.

Signed-off-by: Sasha Levin <sasha.levin@oracle.com>