scsi_ram: a RAM-based SCSI driver
This driver is intended to run as fast as possible, hence the options to
discard writes and reads. It's designed to let us find latency issues
elsewhere in the storage stack (eg filesystem, block layer, scsi layer).

There are a few different options, controlled through module parameters.
The sector size and disc capacity are load-time parameters, but the
parameters affecting performance are tweakable at runtime.

By default, it'll allocate half a gigabyte of RAM to use as a ramdisc;
you can change this with the `capacity' module parameter.

Signed-off-by: Matthew Wilcox <willy@linux.intel.com>
Signed-off-by: Tim Chen <tim.c.chen@linux.intel.com>
[kirill.shutemov@: use vmalloc for scsi_ram_data_array, dummy MODE SENSE handler]
Signed-off-by: Kirill A. Shutemov <kirill.shutemov@linux.intel.com>